ml-audit/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
fairlearn==0.9.0
aif360==0.5.0
shap==0.44.0
joblib==1.3.2
torch==2.1.1
tensorflow==2.15.0
onnx==1.15.0
onnxruntime==1.16.3
httpx==0.25.2
python-dotenv==1.0.0
